The code was converting the aline, but not using it for final checks
When the aline was used, it was added to the list and then converted
from a VT line. However, when the final check came, the original vtline
was used to check if the converted line existed in the output and always
failed. This fixes that by converting the vt line before it is used to
check the results.